requirements.txt
python-dotenv
requests
numpy
pandas
scipy
torch  # PyTorch
jupyterlab
ipywidgets
matplotlib
scikit-learn
chromadb
jupyter-dash
sentencepiece
pyarrow
faiss-cpu

# Pip packages
beautifulsoup4
plotly
bitsandbytes
transformers
sentence-transformers
datasets
accelerate
openai
anthropic
google-generativeai
gradio
gensim
modal
ollama
psutil
setuptools
speedtest-cli
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-community
faiss-cpu
feedparser
twilio
pydub
